chore: avoid nil pointer when sending configuration update
